Move `AppConfig` into `config.rs`
