+++ src/config.rs
@@ -7,8 +7,74 @@ use serde::Deserialize;
 use std::fs;
 use std::path::{Path, PathBuf};
 
+use crate::cli;
 use crate::formatter::OutputFormat;
 use crate::helpers;
+
+/// Represents the final, merged configuration from all sources.
+pub struct AppConfig {
+    pub root: PathBuf,
+    pub full_path: bool,
+    pub prefix: String,
+    pub last_prefix: String,
+    pub child_prefix: String,
+    pub show_all: bool,
+    pub include: Option<String>,
+    pub exclude: Option<String>,
+    pub ignore: Vec<String>,
+    pub directory: bool,
+    pub summary: bool,
+    pub size: bool,
+    pub size_format: helpers::bytes::Format,
+    pub max_depth: Option<usize>,
+    pub format: OutputFormat,
+    pub no_color: bool,
+}
+
+/// Merges settings from the config file and CLI arguments.
+///
+/// CLI arguments take precedence over the config file, which takes precedence over defaults.
+pub fn merge_configs(file: FileConfig, cli: cli::Args) -> AppConfig {
+    AppConfig {
+        // CLI > File > Default
+        root: cli.root.unwrap_or_else(|| PathBuf::from(".")),
+
+        // CLI flags are booleans, so they are always present
+        full_path: cli.full_path || file.full_path.unwrap_or(false),
+        show_all: cli.show_all || file.show_all.unwrap_or(false),
+        directory: cli.directory || file.directory.unwrap_or(false),
+        summary: cli.summary || file.summary.unwrap_or(false),
+        size: cli.size || file.size.unwrap_or(false),
+        no_color: cli.no_color || file.no_color.unwrap_or(false),
+
+        // CLI > File > Default
+        prefix: cli
+            .prefix
+            .or(file.prefix)
+            .unwrap_or_else(|| "├── ".to_string()),
+        last_prefix: cli
+            .last_prefix
+            .or(file.last_prefix)
+            .unwrap_or_else(|| "└── ".to_string()),
+        child_prefix: cli
+            .child_prefix
+            .or(file.child_prefix)
+            .unwrap_or_else(|| "│   ".to_string()),
+
+        // These are optional and can remain None
+        include: cli.include.or(file.include),
+        exclude: cli.exclude.or(file.exclude),
+        max_depth: cli.max_depth.or(file.max_depth),
+
+        // CLI > File > Default
+        ignore: cli.ignore.or(file.ignore).unwrap_or_default(),
+        size_format: cli
+            .size_format
+            .or(file.size_format)
+            .unwrap_or(helpers::bytes::Format::Bytes),
+        format: cli.format.or(file.format).unwrap_or(OutputFormat::Text),
+    }
+}
 
 /// Represents the structure of the configuration file.
 ///
